Finished queries

+++ db/migrate/20160423050012_create_orders.rb
@@ -2,7 +2,7 @@ class CreateOrders < ActiveRecord::Migration
   def change
     create_table :orders do |t|
       t.integer :num_shares
-      t.integer :transaction_fee
+      t.decimal :transaction_fee
       t.boolean :completed
 
       t.belongs_to :account, index: true
@@ -16,7 +16,7 @@ class CreateOrders < ActiveRecord::Migration
       t.string :buy_sell_type
 
       # Hidden Stop Order / Trailing Stop Order
-      t.integer :price_per_share
+      t.decimal :price_per_share
 
       t.timestamps null: false
     end
